chore: add type computer as custom service
